requirements.txt
discord.py
hvac
asqlite @ git+https://github.com/Rapptz/asqlite@main
requests
pyPodcastParser
sqlalchemy[asyncio]
aiosqlite